% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fragments.R
\name{filter_fragments}
\alias{filter_fragments}
\title{Filter DNA fragments in BAM files by length and position}
\usage{
filter_fragments(
  bin_path = "tools/samtools/samtools",
  file = "",
  min_frag_size = 0,
  max_frag_size = 1e+08,
  chr = "",
  start_pos = 1,
  end_pos = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to fastQC executable. Default path tools/samtools/samtools.}

\item{file}{Path to the input file.}

\item{min_frag_size}{Minimum fragment size to keep. Default 1.}

\item{max_frag_size}{Maximum fragment size to keep. Default 10000000.}

\item{chr}{Chromosome to keep. Default None.}

\item{start_pos}{Starting position to search within a chromosome. Default 1}

\item{end_pos}{Last position to search within a chromosome. None}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a BAM file as input and returns a filtered BAM file
by size or/and position, or viceversa.
}
