% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fragments.R
\name{get_fragments_length}
\alias{get_fragments_length}
\title{Extract fragment length from a BAM file.}
\usage{
get_fragments_length(
  bin_path = "~/tools/samtools/samtools",
  bam = "",
  remove_unmapped = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to BAM.}

\item{remove_unmapped}{Removes unmapped reads and/or mates. Only in BAM files. Default false}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a BAM file as input and outputs the fragment length of all reads in a TXT file.
}
