% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fragments.R
\name{merge_bam}
\alias{merge_bam}
\title{Merge BAM files in directory}
\usage{
merge_bam(
  bin_path = "~/tools/samtools/samtools",
  out_bam = "",
  bam_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{out_bam}{Name of merged BAM.}

\item{bam_dir}{Path to directory with BAM files to merge.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a BAM file and merges it with others found within a
directory. This function is still Work In Progress (WIP).
}
