% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fragments.R
\name{plot_fragments_length}
\alias{plot_fragments_length}
\title{Plot DNA fragment length distribution}
\usage{
plot_fragments_length(
  file = "",
  verbose = FALSE,
  min_frag_length = 2,
  max_frag_length = "",
  deviations = 10,
  width_span = 3,
  min_frgl_maximum = 2,
  min_maximum_distance = 10,
  max_maximum_distance = 11,
  vline = ""
)
}
\arguments{
\item{file}{Path to file with fragment length.}

\item{verbose}{Enables progress messages. Default False.}

\item{min_frag_length}{Minimum fragment length to plot. Default 2.}

\item{max_frag_length}{Max fragment length to plot. When not provided it uses the rounded value of MEDIAN + DEVIATIONS*MEDIAN ABSOLUTE DEVIATION as cut point.}

\item{deviations}{MEDIAN + DEVIATIONS*MEDIAN ABSOLUTE DEVIATION. Default 10}

\item{width_span}{Window of width span at which a peak is greater than all other elements around it. Default 3}

\item{min_frgl_maximum}{Minimum fragment length at which to plot maximums peaks. Default 2}

\item{min_maximum_distance}{Minimum distance between local maximum peaks to plot. Default 10}

\item{max_maximum_distance}{Maximum distance between local maximum peaks  to plot. Default 12}

\item{vline}{Fragment length where to draw a vertical line. Default none}
}
\description{
This function takes a file with fragment length as input and plots the length distribution of DNA fragments.
The input file must not have headers and must contain the length of a single fragment per line.
}
