% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{trim_fragments}
\alias{trim_fragments}
\title{Trims reads/fragments in fastqc files}
\usage{
trim_fragments(
  bin_path = "~/tools/bamUtil/bam",
  quality = 33,
  first_base = "",
  last_base = "",
  file_R1 = "",
  file_R2 = ""
)
}
\arguments{
\item{quality}{Min quality of reads to be trimmed.}

\item{first_base}{First base to keep. Trims everything before this base.}

\item{last_base}{Last base to keep. Trims everything after this base.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a fastqc sequence and trims the reads returning a fastqc file with the trimmed reads.
}
