% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fragments.R
\name{get_fragment_length_bed}
\alias{get_fragment_length_bed}
\title{Extract fragment length from a BAM file.}
\usage{
get_fragment_length_bed(
  bin_path = "tools/samtools/samtools",
  bam = "",
  bed = "",
  max_frag_length = 1000,
  mapq = 10,
  threads = 1,
  output_dir = "",
  verbose = FALSE,
  mode = 0
)
}
\arguments{
\item{bin_path}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to BAM file.}

\item{bed}{Path to BED file.}

\item{max_frag_length}{Maximum fragment length to keep. Default 1000.}

\item{mapq}{Minimum MapQ of the reads to keep. Default 10.}

\item{threads}{Number of cores. Default 1}

\item{output_dir}{Directory to output results.}

\item{verbose}{Enables progress messages. Default False.}

\item{mode}{Mode in which to output the fragment length. Default 0. Mode 0: Returns fragment length for R1 and R2 reads. Mode 1: Returns fragment legnth for R1 reads. Mode 2: Returns fragment length for R2 reads.}
}
\description{
This function takes a BAM file and a BED file as an input and outputs the fragment length distributions of all the regions in a TXT file.
}
