% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fragments.R
\name{plot_fragments}
\alias{plot_fragments}
\title{Plot DNA fragment length distribution}
\usage{
plot_fragments(
  bin_path = "tools/samtools/samtools",
  file = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{file}{Path to BAM file.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a BAM file as input and plots the distribution of DNA fragments by their length.
}
